Update chatflow configuration dialog - add starter prompts, analyse chatflow. Rename files.
