+++ packages/ui/src/menu-items/settings.js
@@ -1,14 +1,5 @@
 // assets
-import {
-    IconTrash,
-    IconFileUpload,
-    IconFileExport,
-    IconCopy,
-    IconSearch,
-    IconMessage,
-    IconPictureInPictureOff,
-    IconAdjustmentsHorizontal
-} from '@tabler/icons'
+import { IconTrash, IconFileUpload, IconFileExport, IconCopy, IconMessage, IconAdjustmentsHorizontal } from '@tabler/icons'
 
 // constant
 const icons = {
@@ -16,9 +7,7 @@ const icons = {
     IconFileUpload,
     IconFileExport,
     IconCopy,
-    IconSearch,
     IconMessage,
-    IconPictureInPictureOff,
     IconAdjustmentsHorizontal
 }
 
@@ -29,13 +18,6 @@ const settings = {
     title: '',
     type: 'group',
     children: [
-        {
-            id: 'conversationStarters',
-            title: 'Starter Prompts',
-            type: 'item',
-            url: '',
-            icon: icons.IconPictureInPictureOff
-        },
         {
             id: 'viewMessages',
             title: 'View Messages',
@@ -70,13 +52,6 @@ const settings = {
             type: 'item',
             url: '',
             icon: icons.IconFileExport
-        },
-        {
-            id: 'analyseChatflow',
-            title: 'Analyse Chatflow',
-            type: 'item',
-            url: '',
-            icon: icons.IconSearch
         },
         {
             id: 'deleteChatflow',
